+++ PresentationLayer/src/app/app.module.ts
@@ -1,14 +1,10 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LoginComponent } from './core/login/login.component';
 import { CreditFormComponent } from './core/creditform/creditform.component';
 import { CreditRequestService } from './core/creditform/services/credit-request.service';
-import { UserService } from './shared/user.service';
+import { UserService } from './core/login/login.service';
 import { UrlService } from './shared/url.service';
 import { CreditRequestComponent } from './core/credit-request/credit-request.component';
 import { CreditRequestListComponent } from './core/credit-request-list/credit-request-list.component';
@@ -22,6 +18,9 @@ import { AccountlistComponent } from './core/accountlist/accountlist.component';
 import { AccountService } from './core/account/services/account.service';
 import { CustomerhomeComponent } from './core/customerhome/customerhome.component';
 import { SetUpAccountComponent } from './core/register/set-up-account/set-up-account.component';
+import { LoginComponent } from './core/login/login.component';
+import { HttpClientModule } from '@angular/common/http';
+import { NgModule } from '@angular/core';
