+++ PresentationLayer/src/app/shared/nav-bar/nav-bar.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { UserService } from 'src/app/shared/user.service';
+import { UserService } from 'src/app/core/login/login.service';
 
 @Component({
   selector: 'app-nav-bar',
